Concrete queue classes split into two separate ones
SimpleQueue implements basic queue functionality and SinkingQueue extends it to be able to add new element to full queue
